+++ spec/rspec/core/pending_example_spec.rb
@@ -11,10 +11,21 @@ RSpec.describe "an example" do
     end
   end
 
+  matcher :be_skipped_with do |message|
+    match do |example|
+      example.skipped? && example.metadata[:execution_result][:pending_message] == message
+    end
+
+    failure_message_for_should do |example|
+      "expected: example skipped with #{message.inspect}\n     got: #{example.metadata[:execution_result][:pending_message].inspect}"
+    end
+  end
+
   context "declared pending with metadata" do
     it "uses the value assigned to :pending as the message" do
       group = RSpec::Core::ExampleGroup.describe('group') do
         example "example", :pending => 'just because' do
+          fail
         end
       end
       example = group.examples.first
@@ -25,6 +36,7 @@ RSpec.describe "an example" do
     it "sets the message to 'No reason given' if :pending => true" do
       group = RSpec::Core::ExampleGroup.describe('group') do
         example "example", :pending => true do
+          fail
         end
       end
       example = group.examples.first
@@ -40,7 +52,7 @@ RSpec.describe "an example" do
       end
       example = group.examples.first
       example.run(group.new, double.as_null_object)
-      expect(example).to be_pending_with('Not yet implemented')
+      expect(example).to be_skipped_with('Not yet implemented')
     end
   end
 
@@ -49,11 +61,29 @@ RSpec.describe "an example" do
       group = RSpec::Core::ExampleGroup.describe('group') do
         it "does something" do
           pending
+          fail
         end
       end
       example = group.examples.first
       example.run(group.new, double.as_null_object)
       expect(example).to be_pending_with(RSpec::Core::Pending::NO_REASON_GIVEN)
+    end
+
+    it "fails when the rest of the example passes" do
+      called = false
+      group = RSpec::Core::ExampleGroup.describe('group') do
+        it "does something" do
+          pending
+          called = true
+        end
+      end
+
+      example = group.examples.first
+      example.run(group.new, double.as_null_object)
+      expect(called).to eq(true)
+      result = example.metadata[:execution_result]
+      expect(result[:pending_fixed]).to eq(true)
+      expect(result[:status]).to eq("failed")
     end
   end
 
@@ -173,6 +203,23 @@ RSpec.describe "an example" do
 
       it "passes" do
         expect(run_example.exception).to be_nil
+      end
+
+      it 'indicates it is pending with the given message' do
+        expect(run_example("just because")).to be_pending_with("just because")
+      end
+
+      it 'indicates the pending block was not fixed' do
+        expect(run_example.metadata[:execution_result][:pending_fixed]).to be false
+      end
+    end
+
+    context 'that fails due to both a failed message expectation and a standard failure' do
+      def run_example(*pending_args)
+        super(*pending_args) {
+          expect("foo").to receive("bar")
+          fail
+        }
       end
 
       it 'indicates it is pending with the given message' do